External users with a Microsoft account - problems with encrypted documents
My external user [live account added to Azure AD as guest] cannot open such documents in Office Online, a generic error pops up: Sorry, something went wrong. An error has occurred on the server.  (this happens to my customer as well)
The only way I could have them open these documents is in the desktop app, and this was achievable only by sending an email attachment / copy of the document, rather than generating a sharing link from SharePoint. The problem with this sharing link is that, in a context in which the error is generic, we cannot open in the desktop from it.